% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ldoe.R
\name{calc_proficiency}
\alias{calc_proficiency}
\title{Calculate proficiency rates}
\usage{
calc_proficiency(x, ...)
}
\arguments{
\item{x}{tibble of assessment data}

\item{...}{grouping variables present in data}
}
\description{
Calculate proficiency rates
}
