% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ldoe.R
\name{calc_proficiency_diff}
\alias{calc_proficiency_diff}
\title{Calculate proficiency rates, including difference from previous year}
\usage{
calc_proficiency_diff(x, ..., digits = 0)
}
\arguments{
\item{x}{tibble of assessment data}

\item{...}{grouping variables present in data}

\item{digits}{number of decimals for pretty printing}
}
\description{
Calculate proficiency rates, including difference from previous year
}
