% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ldoe.R
\name{filter_enrolled_cum}
\alias{filter_enrolled_cum}
\title{Filter derivation for records with cumulative enrollment flag}
\usage{
filter_enrolled_cum(x, level)
}
\arguments{
\item{x}{derivation}

\item{level}{"site", "lea", or "state"}
}
\description{
Filter derivation for records with cumulative enrollment flag
}
