% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ldoe.R
\name{filter_enrolled_days}
\alias{filter_enrolled_days}
\title{Filter derivation for records enrolled for a minimum number of days}
\usage{
filter_enrolled_days(x, n_days = 1)
}
\arguments{
\item{x}{derivation}

\item{n_days}{integer}
}
\description{
Filter derivation for records enrolled for a minimum number of days
}
