% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ldoe.R
\name{filter_enrolled_on}
\alias{filter_enrolled_on}
\title{Filter derivation for records enrolled on 10/1 or 2/1}
\usage{
filter_enrolled_on(x, date = "10-01")
}
\arguments{
\item{x}{derivation}

\item{date}{"10-01" or "02-01"}
}
\description{
Filter derivation for records enrolled on 10/1 or 2/1
}
