% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ldoe.R
\name{filter_grades_exclude}
\alias{filter_grades_exclude}
\title{Filter derivation for records NOT in certain grades}
\usage{
filter_grades_exclude(x, grades = c("15", "20"))
}
\arguments{
\item{x}{derivation}

\item{grades}{character vector of LDOE numeric codes}
}
\description{
Filter derivation for records NOT in certain grades
}
