% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata_salesforce_clean.R
\name{getdata_ep_gradelevel}
\alias{getdata_ep_gradelevel}
\title{Get eligibility/priority records by grade level}
\usage{
getdata_ep_gradelevel()
}
\value{
tibble with new field names
}
\description{
Get eligibility/priority records by grade level
}
