% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_augment.R
\name{match_augment}
\alias{match_augment}
\title{Augment match records with additional information}
\usage{
match_augment(m, gradelevels, contactsmatch, choices)
}
\arguments{
\item{m}{tibble of match records}

\item{gradelevels}{tibble of Grade Levels}

\item{contactsmatch}{tibble of Contacts}

\item{choices}{tibble of Application School Rankings}
}
\description{
Augment match records with additional information
}
