% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_match.R
\name{match_calc_program_all}
\alias{match_calc_program_all}
\title{Summarize match outcomes by program}
\usage{
match_calc_program_all(m, schools_waitlist, ...)
}
\arguments{
\item{m}{tibble of match records}

\item{schools_waitlist}{character vector of choice school codes}

\item{...}{grouping variables present in data}
}
\description{
Summarize match outcomes by program
}
