% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_detier.R
\name{match_detier}
\alias{match_detier}
\title{Remove tier structure from match choices}
\usage{
match_detier(dir_business, match)
}
\arguments{
\item{match}{tibble of match records}
}
\value{
tibble of match records, but with a single row for each tiered choice
instead of multiple tiers
}
\description{
Remove tier structure from match choices
}
