% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_notification.R
\name{match_notification}
\alias{match_notification}
\title{Generate upload template for match types}
\usage{
match_notification(dir_business, match, overmatches)
}
\arguments{
\item{dir_business}{character}

\item{match}{tibble of match records}

\item{overmatches}{tibble of siblings assigned over capacity}
}
\description{
Generate upload template for match types
}
