% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_placement.R
\name{match_placement}
\alias{match_placement}
\title{Generate upload template for future-year placements}
\usage{
match_placement(dir_business, match, overmatches)
}
\arguments{
\item{dir_business}{character}

\item{match}{tibble of match records}

\item{overmatches}{tibble of siblings assigned over capacity}
}
\description{
Generate upload template for future-year placements
}
