% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_process}
\alias{match_process}
\title{Process match file}
\usage{
match_process(run, dir_in = "in", dir_out = "out", use_cache = FALSE)
}
\arguments{
\item{run}{integer}

\item{dir_in}{character}

\item{dir_out}{character}

\item{use_cache}{logical}
}
\description{
Process match file
}
