% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_summary.R
\name{match_summary_program}
\alias{match_summary_program}
\title{Summarize program-grade match outcomes}
\usage{
match_summary_program(m, schools_waitlist, ...)
}
\arguments{
\item{m}{tibble of match records}

\item{schools_waitlist}{character vector of choice school codes}

\item{...}{additional grouping variables, like program and grade}
}
\description{
Summarize program-grade match outcomes
}
