% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_summary.R
\name{match_summary_student}
\alias{match_summary_student}
\title{Summarize student-level match outcomes}
\usage{
match_summary_student(m, schools_waitlist, ...)
}
\arguments{
\item{m}{tibble of match records}

\item{schools_waitlist}{character vector of choice school codes}

\item{...}{additional student-level grouping variables, like grade}
}
\description{
Summarize student-level match outcomes
}
