% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_util.R
\name{matchtypes_salesforce}
\alias{matchtypes_salesforce}
\title{Match outcome categories used in Salesforce}
\usage{
matchtypes_salesforce()
}
\value{
tibble
}
\description{
Match outcome categories used in Salesforce
}
