% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_util.R
\name{schools_net}
\alias{schools_net}
\title{Programs requiring 8th grade applicants to be at least 15 by 9/30}
\usage{
schools_net()
}
\value{
character vector
}
\description{
Programs requiring 8th grade applicants to be at least 15 by 9/30
}
