% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_augment.R
\name{match_augment}
\alias{match_augment}
\title{Augment lottery records with grade level and contact information}
\usage{
match_augment(x, gradelevels, contactsmatch)
}
\arguments{
\item{x}{tibble of lottery records}

\item{gradelevels}{tibble of Grade Level records}

\item{contactsmatch}{tibble of Contact records}
}
\description{
Augment lottery records with grade level and contact information
}
