% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_augment.R
\name{match_augment}
\alias{match_augment}
\title{Augment match records with additional information}
\usage{
match_augment(m, gradelevels, contactsmatch, choices, expulsions)
}
\arguments{
\item{m}{tibble of match records}

\item{gradelevels}{tibble of grade levels}

\item{contactsmatch}{tibble of contacts}

\item{choices}{tibble of application school rankings}

\item{expulsions}{tibble of expulsions}
}
\description{
Augment match records with additional information
}
